test: run eslint v8 test in the e2e job
